NeoEloquentServiceProvider now works for Laravel 5.3 / removed Framework/Laravel
